+++ docs/CHANGELOG.md
@@ -45,7 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix incorrect ACLs for flaw drafts (OSIDB-2263)
 - Fix workflow rejection endpoint (OSIDB-2456)
 - Fix FlawReference article count validation (OSIDB-2651)
-- Fix not being able to set CVE ID to an empty string through the API (OSIDB-270)
+- Fix not being able to set CVE ID to an empty string through the API (OSIDB-2702)
 - Comments not properly updating when syncing from Bugzilla (OSIDB-1385)
 
 ### Removed